inventory management is partially functional.
Did a lot of work to get item updates working and I have some of the
code to add new items in place.  Had to fix a few issues in some of the
SQL code and made some changes to the database (for example added an
"author" column).  Also when I added some new items found that they were
getting excluded because we were doing an INNER JOIN with the ratings
table (but they didn't have any ratings).  Changed this to an OUTER
JOIN.  And various other things.